Perforce: Do not wait for process if not set up
The process could be not set up correctly or at all if
we exit the start function early.
Avoids a soft assert.

+++ src/plugins/perforce/perforcechecker.cpp
@@ -26,8 +26,10 @@ PerforceChecker::PerforceChecker(QObject *parent) : QObject(parent)
 
 PerforceChecker::~PerforceChecker()
 {
-    m_process.kill();
-    m_process.waitForFinished();
+    if (m_process.isRunning()) {
+        m_process.kill();
+        m_process.waitForFinished();
+    }
     resetOverrideCursor();
 }
